+++ src/protocols/types.rs
@@ -29,7 +29,6 @@ pub enum Protocol {
 }
 
 /// All response types
-#[cfg_attr(feature = "serde", derive(Serialize, Deserialize))]
 #[derive(Debug, Clone, PartialEq)]
 pub enum GenericResponse<'a> {
     GameSpy(gamespy::VersionedResponse<'a>),
@@ -45,7 +44,6 @@ pub enum GenericResponse<'a> {
 }
 
 /// All player types
-#[cfg_attr(feature = "serde", derive(Serialize, Deserialize))]
 #[derive(Debug, Clone, PartialEq)]
 pub enum GenericPlayer<'a> {
     Valve(&'a valve::ServerPlayer),